GameWindowクラスをrecordで実装からクラスで実装するように変更して,バグを解決できました。
